requirements.txt
yt-dlp
openai-whisper
torch
setuptools-rust
google-generativeai
python-dotenv
